Setup attach overrride
